Fix: Revert extraction of constraints
