--- test/Unit/DataProvider.php
@@ -1,64 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-/**
- * Copyright (c) 2017 Andreas Möller.
- *
- * For the full copyright and license information, please view
- * the LICENSE file that was distributed with this source code.
- *
- * @link https://github.com/localheinz/test-util
- */
-
-namespace Localheinz\Test\Util\Test\Unit;
-
-use Localheinz\Test\Util\Test\Fixture;
-
-final class DataProvider
-{
-    public function providerNotClass(): \Generator
-    {
-        $classyNames = [
-            'class-non-existent' => __NAMESPACE__ . '\NonExistentClass',
-            'interface' => Fixture\NotClass\ExampleInterface::class,
-            'trait' => Fixture\NotClass\ExampleTrait::class,
-        ];
-
-        foreach ($classyNames as $key => $classyName) {
-            yield $key => [
-                $classyName,
-            ];
-        }
-    }
-
-    public function providerNotInterface(): \Generator
-    {
-        $interfaceNames = [
-            'class' => Fixture\NotInterface\ExampleClass::class,
-            'interface-non-existent' => __NAMESPACE__ . '\NonExistentInterface',
-            'trait' => Fixture\NotInterface\ExampleTrait::class,
-        ];
-
-        foreach ($interfaceNames as $key => $interfaceName) {
-            yield $key => [
-                $interfaceName,
-            ];
-        }
-    }
-
-    public function providerNotTrait(): \Generator
-    {
-        $traitNames = [
-            'class' => Fixture\NotTrait\ExampleClass::class,
-            'interface' => Fixture\NotTrait\ExampleInterface::class,
-            'trait-non-existent' => __NAMESPACE__ . '\NonExistentTrait',
-        ];
-
-        foreach ($traitNames as $key => $traitName) {
-            yield $key => [
-                $traitName,
-            ];
-        }
-    }
-}
